Extract DatePicker to own component
